back/ai/requirements.txt
fastapi
uvicorn
python-dotenv
asyncio
openai